feat(phase-7): Generation u32 cascade (Step 0+1)
Narrow Generation alias from usize to u32 to enable 16-byte Entity
and EntityInland layouts on the upcoming fast random-access path.

Step 0 (boyko_utils):
- primitives.rs: pub type Generation = usize -> u32.
- sparse_slot_map.rs: push_dense(.., generation: Generation).

Step 1 (boyko_ecs):
- entity.rs: Entity::generation: usize -> u32; Entity::new,
  generation(), From<Slot>/From<Entity> for Slot flow u32 directly.
- entity_master.rs:84, ecs_master.rs:170: entity.generation() as
  usize bridge cast at the old EntityInland::new boundary
  (transient, removed by Step 9).

Wrap window per D2: 2^32 deallocs/slot >= 13.6 years at 10/sec.
EcsMaster transient as usize cast disappears when old EntityInland
is deleted in Step 9.

Acceptance: cargo check --all-targets clean, cargo clippy clean,
cargo test --workspace 258/258 pass.

+++ crates/boyko_ecs/src/ecs/core/ecs_master/ecs_master.rs
@@ -167,7 +167,7 @@ impl EcsMaster {
 
         // Allocate a new entity — only reached if the guard passed.
         let entity = self.entity_master.allocate_entity();
-        let generation = entity.generation();
+        let generation = entity.generation() as usize;
 
         // Create EntityInland with initial values.
         let mut entity_inland = EntityInland::new(
